separate KeyUtil from KeyPairUtil not to load android store code when unit test
